+++ obj/src/persistence/IdentifiableSqlitePersistence.js
@@ -88,6 +88,10 @@ class IdentifiableSqlitePersistence extends SqlitePersistence_1.SqlitePersistenc
      */
     constructor(tableName) {
         super(tableName);
+        /**
+         * Flag to turn on automated string ID generation
+         */
+        this._autoGenerateId = true;
         if (tableName == null) {
             throw new Error("Table name could not be null");
         }
@@ -174,7 +178,7 @@ class IdentifiableSqlitePersistence extends SqlitePersistence_1.SqlitePersistenc
             }
             // Assign unique id
             let newItem = item;
-            if (newItem.id == null) {
+            if (newItem.id == null && this._autoGenerateId) {
                 newItem = Object.assign({}, newItem);
                 newItem.id = item.id || pip_services3_commons_nodex_1.IdGenerator.nextLong();
             }
@@ -195,7 +199,7 @@ class IdentifiableSqlitePersistence extends SqlitePersistence_1.SqlitePersistenc
                 return null;
             }
             // Assign unique id
-            if (item.id == null) {
+            if (item.id == null && this._autoGenerateId) {
                 item = Object.assign({}, item);
                 item.id = pip_services3_commons_nodex_1.IdGenerator.nextLong();
             }