+++ src/persistence/IdentifiableSqlitePersistence.ts
@@ -79,6 +79,10 @@ import { SqlitePersistence } from './SqlitePersistence';
  */
 export class IdentifiableSqlitePersistence<T extends IIdentifiable<K>, K> extends SqlitePersistence<T>
     implements IWriter<T, K>, IGetter<T, K>, ISetter<T> {
+    /**
+     * Flag to turn on automated string ID generation
+     */
+    protected _autoGenerateId: boolean = true;
 
     /**
      * Creates a new instance of the persistence component.
@@ -177,7 +181,7 @@ export class IdentifiableSqlitePersistence<T extends IIdentifiable<K>, K> extend
 
         // Assign unique id
         let newItem: any = item;
-        if (newItem.id == null) {
+        if (newItem.id == null && this._autoGenerateId) {
             newItem = Object.assign({}, newItem);
             newItem.id = item.id || IdGenerator.nextLong();
         }
@@ -199,7 +203,7 @@ export class IdentifiableSqlitePersistence<T extends IIdentifiable<K>, K> extend
         }
 
         // Assign unique id
-        if (item.id == null) {
+        if (item.id == null && this._autoGenerateId) {
             item = Object.assign({}, item);
             item.id = <any>IdGenerator.nextLong();
         }